svara_inference.py
import torch
from nemo.collections.tts.models import SpectrogramGeneratorModel, VocoderModel

# Note: You must place your Svara weights in a /models/svara folder
def generate_svara_audio(text, output_path="comfyui/input/voice.wav"):
    # Load Svara-v1 model logic here
    # 80/20 Mix logic is handled by passing the raw text to Svara
    # which supports code-switching naturally.
    print(f"Generating audio for: {text}")
    # Save to ComfyUI input folder so the workflow can see it
    pass
